Add `hostNetwork: true` for compat with Workload Identity
The daemonset relies on reading the preemption endpoint of the node's
metadata server.

When using GKE Workload Identity, the daemonset is
instead talking to gke-metadata-server, which doesn't expose the
preemption endpoint.

This can be bypassed by running the daemonset in the host network
namespace, so that it still talks to the node metadata server.